# libs/ocr_common/ocr_common/schemas.py
"""
Skema response yang sama untuk semua service: envelope sukses, error (+ contoh), job, health.

Teks di dalam Field(...) muncul apa adanya di Swagger (/docs) dan openapi.yaml yang dibaca tim lain
(gateway / orkestrasi), jadi ditulis dalam bahasa Inggris dan selalu membawa contoh.
"""

from typing import Any, Literal

from pydantic import BaseModel, Field

from ocr_common.envelope import envelope

REQUEST_ID_EXAMPLE = "REQ_9cb01af2-493d-446d-b191-af120333f6d0"

Stage = Literal["OCR", "STRUCTURING", "SCORING"]
JobState = Literal["PROCESSING", "DONE", "FAILED"]


class SuccessEnvelope(BaseModel):
    """Bagian envelope yang sama untuk semua response sukses; subclass hanya menentukan `data`."""

    status_code: int = Field(200, description="Same as the HTTP status code", examples=[200])
    status_desc: str = Field("OK", description="Reason phrase of `status_code`", examples=["OK"])
    message: str = Field("Success", description="Human-readable outcome", examples=["Success"])
    errors: None = Field(None, description="Always null on success")
    request_id: str | None = Field(
        None,
        description=(
            "The request_id this response belongs to: the one sent in the form / body / path, otherwise the "
            "`X-Request-ID` header, otherwise one generated by the service"
        ),
        examples=[REQUEST_ID_EXAMPLE],
    )


class ErrorResponse(BaseModel):
    status_code: int = Field(..., description="Same as the HTTP status code", examples=[400])
    status_desc: str = Field(..., description="Reason phrase of `status_code`", examples=["Bad Request"])
    message: str = Field(..., description="Human-readable cause; safe to log", examples=["Uploaded file is empty"])
    data: None = Field(None, description="Always null on error")
    errors: str = Field(
        ...,
        description=(
            "Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message names a different "
            "field every time); for every other status it repeats `message`"
        ),
        examples=["Uploaded file is empty"],
    )
    request_id: str | None = Field(
        None, description="Present when the error is tied to a known request_id", examples=[REQUEST_ID_EXAMPLE]
    )


class JobAccepted(BaseModel):
    request_id: str = Field(..., description="Echo of the submitted request_id", examples=[REQUEST_ID_EXAMPLE])
    stage: Stage = Field(..., description="Pipeline stage that accepted the job", examples=["OCR"])
    status: JobState = Field(
        ...,
        description=(
            "`PROCESSING` for a newly accepted job. When `duplicate` is true: the current status of the job "
            "that already exists for this request_id"
        ),
        examples=["PROCESSING"],
    )
    duplicate: bool = Field(
        False,
        description=(
            "true: this request_id was already accepted before and the work is NOT run again (idempotent). "
            "Exception: a job whose status is `FAILED` is re-claimed and run again, and answers `duplicate: false`"
        ),
        examples=[False],
    )


class JobAcceptedResponse(SuccessEnvelope):
    status_code: int = Field(202, description="Same as the HTTP status code", examples=[202])
    status_desc: str = Field("Accepted", description="Reason phrase of `status_code`", examples=["Accepted"])
    message: str = Field("Accepted", description="Human-readable outcome", examples=["Accepted"])
    data: JobAccepted


class JobStatusBase(BaseModel):
    """Status satu tahap untuk satu request_id. Tiap service menurunkannya dengan `result` yang bertipe."""

    request_id: str = Field(..., description="request_id of the pipeline run", examples=[REQUEST_ID_EXAMPLE])
    stage: Stage = Field(..., description="Pipeline stage this job belongs to", examples=["OCR"])
    status: JobState = Field(
        ...,
        description="`PROCESSING` -> `DONE` | `FAILED`. A `FAILED` job may go back to `PROCESSING` when resubmitted",
        examples=["DONE"],
    )
    error_message: str | None = Field(
        None,
        description="Why the stage failed; null unless `status` is `FAILED`. Same text as the FAILED callback",
        examples=[None],
    )
    created_at: str = Field(
        ..., description="When the job was first accepted (ISO 8601, UTC)", examples=["2026-09-18T04:00:00+00:00"]
    )
    updated_at: str = Field(
        ..., description="Last status change (ISO 8601, UTC)", examples=["2026-09-18T04:00:01+00:00"]
    )


class HealthResponse(BaseModel):
    status: str = Field(..., description="Always `healthy` while the process is alive", examples=["healthy"])
    version: str = Field(..., description="Service version", examples=["1.0.0"])
    device: str = Field(..., description="Where this service computes", examples=["cpu"])
    backends: dict[str, str] = Field(
        ...,
        description="Active implementation per concern (model backend, storage). Useful to confirm a deployment",
        examples=[{"ekstraksi": "paddle", "storage": "postgres"}],
    )


class ReadyResponse(BaseModel):
    status: Literal["ready", "not_ready"] = Field(
        ..., description="`ready` (HTTP 200) when every check is `ok`, else `not_ready` (HTTP 503)", examples=["ready"]
    )
    checks: dict[str, Literal["ok", "failed"]] = Field(
        ...,
        description="One entry per REQUIRED dependency of this service; empty when it has none",
        examples=[{"database": "ok"}],
    )


def error(
    status_code: int,
    description: str,
    message: str,
    *,
    request_id: str | None = None,
    errors: str | None = None,
) -> dict:
    """Satu entri `responses` lengkap dengan contohnya sendiri.

    Tanpa contoh per status, /docs memakai example bawaan ErrorResponse untuk
    semua kode: 401 pun tampil sebagai 400, dan pembacanya belajar kode yang
    salah."""
    return {
        "model": ErrorResponse,
        "description": description,
        "content": {
            "application/json": {
                "example": envelope(status_code, message, None, request_id, errors=errors or message),
            }
        },
    }


def success_examples(description: str, **named: tuple[str, Any]) -> dict:
    """Entri `responses` 2xx dengan contoh bernama: `nama=(ringkasan, body lengkap)`.

    Satu contoh per skenario yang harus ditangani pemanggil (mis. job baru vs duplikat), bukan satu
    contoh rakitan dari `examples` tiap field: yang terakhir tidak pernah menunjukkan kombinasi nyata."""
    return {
        "description": description,
        "content": {
            "application/json": {
                "examples": {name: {"summary": summary, "value": value} for name, (summary, value) in named.items()}
            }
        },
    }


# Entri 401 yang sama di semua router; diimpor supaya kalimatnya satu sumber.
UNAUTHORIZED = error(401, "Missing or invalid X-API-Key", "Invalid or missing API key")