# services/ekstraksi/openapi.yaml
openapi: 3.1.0
info:
  title: OCR NPWP API
  description: 'OCR service for Indonesian NPWP (tax ID card) documents; ServiceOCR in the pipeline. **Async
    pipeline:** the orchestrator POSTs /v1/ekstraksi/jobs after guardrails passed and gets 202; this service
    runs OCR in the background, stores the result, POSTs a stage callback to the orchestrator, and hands
    the job to the structuring service. **Legacy contract:** generate-request-id -> extract-ocr -> get-ocr-result
    runs the whole chain synchronously and stays until the orchestrator has moved to the async flow. The
    raw OCR step is also exposed as /v1/ekstraksi/extract. All endpoints except /health require an X-API-Key
    header.


    ## Conventions (the same for every nilam-ocr service)


    **Authentication.** Every endpoint except `/health` and `/ready` requires the header `X-API-Key`.

    A missing or wrong key answers `401`. The services are reachable only from inside the cluster.


    **Envelope.** Every JSON response, success or error, has the same shape:

    `{status_code, status_desc, message, data, errors, request_id}`. `status_code` always equals the HTTP

    status. On success `errors` is null; on error `data` is null.


    **Errors.** `errors` is machine-readable: `VALIDATION_ERROR` for `422`, otherwise the same text as

    `message`. `400` = the request or the document is unusable (do not retry unchanged). `401` = API key.

...

...

...


...
    no

...

    without it the service generates one.


...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/ekstraksi/extract:
    post:
      tags:
      - Ekstraksi
      summary: Raw OCR of a document, synchronous (no job, no callback)
      description: Runs the OCR engine and returns raw text blocks with confidence and bounding boxes.
        Send the image as `file` or `file_url`; exactly one of the two.
      operationId: extractText
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_extractText'
      responses:
        '200':
          description: Text lines found in the document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractResponse'
              examples:
                npwp_card:
                  summary: An NPWP card
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      engine: paddle
                      model: PP-OCRv6_medium_det+PP-OCRv6_medium_rec
                      elapsed_ms: 412.5
                      full_text: 'DIREKTORAT JENDERAL PAJAK

                        NPWP : 12.345.678.9-012.345

                        BUDI SANTOSO'
                      blocks:
                      - text: DIREKTORAT JENDERAL PAJAK
                        confidence: 0.9985
                        bbox:
                          x1: 175
                          y1: 247
                          x2: 887
                          y2: 308
                        page: 0
                      - text: 'NPWP : 12.345.678.9-012.345'
                        confidence: 0.9992
                        bbox:
                          x1: 35
                          y1: 389
                          x2: 637
                          y2: 437
                        page: 0
                      - text: BUDI SANTOSO
                        confidence: 0.9773
                        bbox:
                          x1: 33
                          y1: 480
                          x2: 483
                          y2: 523
                        page: 0
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '400':
          description: Bad file (empty, too large, unsupported type) or bad intake
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 400
                status_desc: Bad Request
                message: Uploaded file is empty
                errors: Uploaded file is empty
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
        '500':
          description: OCR engine failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 500
                status_desc: Internal Server Error
                message: 'ekstraksi OCR model error (500): error: OpenCV ...'
                errors: 'ekstraksi OCR model error (500): error: OpenCV ...'
        '503':
          description: OCR model unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 503
                status_desc: Service Unavailable
                message: ekstraksi OCR model is unavailable
                errors: ekstraksi OCR model is unavailable
        '504':
          description: OCR model did not answer in time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 504
                status_desc: Gateway Timeout
                message: ekstraksi OCR model timed out after 30.0s
                errors: ekstraksi OCR model timed out after 30.0s
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'body.file: Expected UploadFile, received: str'
                errors: VALIDATION_ERROR
      security:
      - APIKeyHeader: []
...
components:
  schemas:
...
    Body_extractText:
      properties:
        file:
          anyOf:
          - type: string
            format: binary
          - type: string
          - type: 'null'
          title: File
          description: Document image (JPEG/PNG/PDF). Omit when sending file_url.
        file_url:
          anyOf:
          - type: string
          - type: 'null'
          title: File Url
          description: URL this service fetches the document image from (e.g. a presigned MinIO GET).
            Omit when uploading file.
      type: object
      title: Body_extractText
...
    BoundingBox:
      properties:
        x1:
          type: integer
          title: X1
          description: Left
          examples:
          - 190
        y1:
          type: integer
          title: Y1
          description: Top
          examples:
          - 320
        x2:
          type: integer
          title: X2
          description: Right
          examples:
          - 760
        y2:
          type: integer
          title: Y2
          description: Bottom
          examples:
          - 366
      type: object
      required:
      - x1
      - y1
      - x2
      - y2
      title: BoundingBox
...
    ErrorResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          examples:
          - 400
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          examples:
          - Bad Request
        message:
          type: string
          title: Message
          description: Human-readable cause; safe to log
          examples:
          - Uploaded file is empty
        data:
          type: 'null'
          title: Data
          description: Always null on error
        errors:
          type: string
          title: Errors
          description: Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message
            names a different field every time); for every other status it repeats `message`
          examples:
          - Uploaded file is empty
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Present when the error is tied to a known request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - status_code
      - status_desc
      - message
      - errors
      title: ErrorResponse
...
    ExtractResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/OcrResult'
      type: object
      required:
      - data
      title: ExtractResponse
...
    OcrResult:
      properties:
        engine:
          type: string
          title: Engine
          description: OCR backend that read the document (`EKSTRAKSI_BACKEND`)
          examples:
          - paddle
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: Model identity reported by the OCR model service; null when it does not report
            one
          examples:
          - PP-OCRv6_medium_det+PP-OCRv6_medium_rec
        elapsed_ms:
          type: number
          title: Elapsed Ms
          description: Time spent in the OCR model, milliseconds
          examples:
          - 412.5
        full_text:
          type: string
          title: Full Text
          description: All blocks joined by newline, in reading order
          examples:
          - 'DIREKTORAT JENDERAL PAJAK

            NPWP : 12.345.678.9-012.345

            BUDI SANTOSO'
        blocks:
          items:
            $ref: '#/components/schemas/TextBlock'
          type: array
          title: Blocks
          description: Text lines in reading order (top to bottom)
      type: object
      required:
      - engine
      - elapsed_ms
      - full_text
      - blocks
      title: OcrResult
...
    TextBlock:
      properties:
        text:
          type: string
          title: Text
          description: One OCR text line, as read
          examples:
          - 'NPWP : 12.345.678.9-012.345'
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          description: Recognition score of this line
          examples:
          - 0.9992
        bbox:
          anyOf:
          - $ref: '#/components/schemas/BoundingBox'
          - type: 'null'
          description: Upright box around the line, in pixels of the ORIENTATION-CORRECTED image the OCR
            model worked on. The model straightens and unwarps the card first, so these are NOT coordinates
            in the uploaded photo
        page:
          type: integer
          minimum: 0.0
          title: Page
          description: 0-based page number (multi-page PDF)
          default: 0
          examples:
          - 0
      type: object
      required:
      - text
      - confidence
      title: TextBlock
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: API key for this service
      in: header
      name: X-API-Key
...